send-email: fix bug resulting in increased message number if a message is edited
In case a message is edited before it is sent, its message number gets
increased by 1, and so does its order in the message id. The cause of
this bug was that when a person attempts to edit the message, the whole
sub process_file gets terminated, and the user is asked to edit the message.
After necessary edits are done, the whole sub process_file is executed again.
The way sub process_file is designed, every time is runs, it increases the
$message_num variable by 1. The reason for this was that the function ran
again everytime a next message was sent in a thread, and thus we need to
increase the message number for that message. In case a user edits the message,
there is no check for the same and the new message gets treated as a subsequent
message of a thread, therefore increasing its message number by one. This
breaks the shallow thread logic which relies on $message_num being 1 for the
first message, and it gets changed in case the user edits the first message.

So, upon scanning the whole code, there are two significant variables at play
here. First is $message_num, responsible for the message number and second
is $message_id_serial, responsible for showing the message number in the
Message-ID header. So, whenever we edit a message, lets just decrease them
by 1, so that when the whole process to compose and send the message starts,
these variables increase by 1 again, thus get set to the original values for
that message.

We also are doing the same thing in case the user chooses to not send a message
out of many messages in a thread. By doing so, we will simply decrease these
variables by 1 for further messages, thus ensuring the whole thread doesn't
break.

Signed-off-by: Aditya Garg <gargaditya08@live.com>